Avanaces de matriz legal

+++ app/Http/Controllers/LegalAspects/LegalMatrix/InterestController.php
@@ -45,7 +45,10 @@ class InterestController extends Controller
     */
     public function data(Request $request)
     {
-        $interests = Interest::select('*');
+        if ($request->has('custom'))
+            $interests = Interest::company()->select('*');
+        else
+            $interests = Interest::system()->select('*');
 
         return Vuetable::of($interests)
                     ->make();
@@ -60,9 +63,17 @@ class InterestController extends Controller
     public function store(InterestRequest $request)
     {
         $interest = new Interest($request->all());
+
+        if ($request->custom == 'true')
+            $interest->company_id = Session::get('company_id');
         
-        if(!$interest->save()){
+        if (!$interest->save())
             return $this->respondHttp500();
+
+        if ($request->custom == 'true')
+        {
+            $company = Company::find(Session::get('company_id'));
+            $company->interests()->attach($interest->id);
         }
 
         return $this->respondHttp200([
@@ -81,6 +92,7 @@ class InterestController extends Controller
         try
         {
             $interest = Interest::findOrFail($id);
+            $interest->custom = $interest->company_id ? true : false;  
 
             return $this->respondHttp200([
                 'data' => $interest,
@@ -118,12 +130,17 @@ class InterestController extends Controller
      */
     public function destroy(Interest $interest)
     {
-        if(!$interest->delete())
+        if (COUNT($interest->articles) > 0)
+        {
+            return $this->respondWithError('No se puede eliminar el interés porque hay registros asociados a el');
+        }
+
+        if (!$interest->delete())
         {
             return $this->respondHttp500();
         }
 
-        SyncQualificationsCompaniesJob::dispatch();
+        //SyncQualificationsCompaniesJob::dispatch();
         
         return $this->respondHttp200([
             'message' => 'Se elimino el intereses'
@@ -150,7 +167,7 @@ class InterestController extends Controller
         try
         {
             $company = Company::find(Session::get('company_id'));
-            $values = $company->interests()->pluck('sau_lm_interests.id');
+            $values = $company->interests()->system()->pluck('sau_lm_interests.id');
             
             return $this->respondHttp200([
                 'data' => [
@@ -164,38 +181,14 @@ class InterestController extends Controller
         }
     }
 
-    /**
-     * Returns an array for a select type input
-     *
-     * @param Request $request
-     * @return Array
-     */
-
-    public function multiselect(Request $request)
+    public function multiselectSystem(Request $request)
     {
-        if($request->has('keyword'))
-        {
-            $keyword = "%{$request->keyword}%";
-            $interests = Interest::select("id", "name")
-                ->where(function ($query) use ($keyword) {
-                    $query->orWhere('name', 'like', $keyword);
-                })
-                ->take(30)->pluck('id', 'name');
+        return $this->multiselect($request, 'system');
+    }
 
-            return $this->respondHttp200([
-                'options' => $this->multiSelectFormat($interests)
-            ]);
-        }
-        else
-        {
-            $interests = Interest::select(
-                'sau_lm_interests.id as id',
-                'sau_lm_interests.name as name'
-            )
-            ->pluck('id', 'name');
-        
-            return $this->radioFormat($interests);
-        }
+    public function multiselectCompany(Request $request)
+    {
+        return $this->multiselect($request, 'company');
     }
 
     /**
@@ -205,14 +198,13 @@ class InterestController extends Controller
      * @return Array
      */
 
-    public function multiselectCompany(Request $request)
+    public function multiselect(Request $request, $scope = 'alls')
     {
         if($request->has('keyword'))
         {
             $keyword = "%{$request->keyword}%";
             $interests = Interest::select("id", "name")
-                ->join('sau_lm_company_interest', 'sau_lm_company_interest.interest_id', 'sau_lm_interests.id')
-                ->where('sau_lm_company_interest.company_id', Session::get('company_id'))
+                ->$scope()
                 ->where(function ($query) use ($keyword) {
                     $query->orWhere('name', 'like', $keyword);
                 })
@@ -228,8 +220,7 @@ class InterestController extends Controller
                 'sau_lm_interests.id as id',
                 'sau_lm_interests.name as name'
             )
-            ->join('sau_lm_company_interest', 'sau_lm_company_interest.interest_id', 'sau_lm_interests.id')
-            ->where('sau_lm_company_interest.company_id', Session::get('company_id'))
+            ->$scope()
             ->pluck('id', 'name');
         
             return $this->radioFormat($interests);
